+++ themes/hexo/components/TocDrawerButton.js
@@ -10,10 +10,10 @@ import CONFIG_HEXO from '../config_hexo'
  * @constructor
  */
 const TocDrawerButton = (props) => {
+  const { locale } = useGlobal()
   if (!CONFIG_HEXO.WIDGET_TOC) {
     return <></>
   }
-  const { locale } = useGlobal()
   return (<div onClick={props.onClick} className='py-2 px-3 cursor-pointer transform duration-200 flex justify-center items-center w-7 h-7 text-center' title={locale.POST.TOP} >
     <i className='fas fa-list-ol text-xs'/>
   </div>)
