+++ themes/medium/components/Catalog.js
@@ -2,7 +2,7 @@ import React, { useRef } from 'react'
 import throttle from 'lodash.throttle'
 import { uuidToId } from 'notion-utils'
 import Progress from './Progress'
-import JumpToTopButton from './JumpToTopButton'
+
 /**
  * 目录导航组件
  * @param toc
@@ -85,7 +85,6 @@ const Catalog = ({ toc }) => {
         })}
       </nav>
     </div>
-    <JumpToTopButton className='text-gray-400 hover:text-green-500 hover:bg-gray-100 py-1 duration-200' />
   </div>
 }
